% Generated by roxygen2 (4.0.0): do not edit by hand
\name{make_null}
\alias{make_null}
\title{make_null}
\usage{
make_null(mat, iter = 100, method = "rc")
}
\arguments{
\item{mat}{a binary matrix}

\item{iter}{the number of null matrices to create. Defaults to 100}

\item{method}{which null model you want to use. Defaults to method = "rc}
}
\description{
This function generates a null distribution of matrices using one of 5 methods
}

