% Generated by roxygen2 (4.0.0): do not edit by hand
\name{permutes_prob}
\alias{permutes_prob}
\title{permutes_prob
This function generates a null distribution of matrices by maintaining row and column probabilities}
\usage{
permutes_prob(mat, iter = 100)
}
\arguments{
\item{mat}{a binary matrix}

\item{iter}{the number of null matrices to create. Defaults to 100}
}
\description{
permutes_prob
This function generates a null distribution of matrices by maintaining row and column probabilities
}

