% Generated by roxygen2 (4.0.0): do not edit by hand
\name{permutes_r}
\alias{permutes_r}
\title{permutes_r
This function generates a null distribution of matrices by maintaining row sums}
\usage{
permutes_r(mat, iter = 100)
}
\arguments{
\item{mat}{a binary matrix}

\item{iter}{the number of null matrices to create. Defaults to 100}
}
\description{
permutes_r
This function generates a null distribution of matrices by maintaining row sums
}

