% Generated by roxygen2 (4.0.0): do not edit by hand
\name{permutes_rc}
\alias{permutes_rc}
\title{permutes_rc
This function generates a null distribution of matrices by maintaining row and column sums}
\usage{
permutes_rc(mat, iter = 100)
}
\arguments{
\item{mat}{a binary matrix}

\item{iter}{the number of null matrices to create. Defaults to 100}
}
\description{
permutes_rc
This function generates a null distribution of matrices by maintaining row and column sums
}

