% Generated by roxygen2 (4.0.0): do not edit by hand
\name{permutes_rcd}
\alias{permutes_rcd}
\title{permutes_rcd
This function generates a null distribution of matrices by maintaining row and column sums as well as the diagonal}
\usage{
permutes_rcd(mat, iter = 100)
}
\arguments{
\item{mat}{a binary matrix}

\item{iter}{the number of null matrices to create. Defaults to 100}
}
\description{
permutes_rcd
This function generates a null distribution of matrices by maintaining row and column sums as well as the diagonal
}

