% Generated by roxygen2 (4.0.0): do not edit by hand
\name{permutes_c}
\alias{permutes_c}
\title{permutes_c}
\usage{
permutes_c(mat, iter = 100)
}
\arguments{
\item{mat}{a binary matrix}

\item{iter}{the number of null matrices to create. Defaults to 100}
}
\description{
This function generates a null distribution of matrices by maintaining column sums
}
\seealso{
Other null.models: \code{\link{permutes_prob}};
  \code{\link{permutes_rcd}}; \code{\link{permutes_rc}};
  \code{\link{permutes_r}}
}

