% Generated by roxygen2 (4.0.0): do not edit by hand
\name{permutes_rcd}
\alias{permutes_rcd}
\title{permutes_rcd}
\usage{
permutes_rcd(mat, iter = 100)
}
\arguments{
\item{mat}{a binary matrix}

\item{iter}{the number of null matrices to create. Defaults to 100}
}
\description{
This function generates a null distribution of matrices by maintaining row and column sums as well as the diagonal
}
\seealso{
Other null.models: \code{\link{permutes_c}};
  \code{\link{permutes_prob}}; \code{\link{permutes_rc}};
  \code{\link{permutes_r}}
}

